Implementando outras exceções

+++ src/main/java/com/devjulio/dscommerce/services/ProductService.java
@@ -3,10 +3,14 @@ package com.devjulio.dscommerce.services;
 import com.devjulio.dscommerce.DTO.ProductDTO;
 import com.devjulio.dscommerce.entities.Product;
 import com.devjulio.dscommerce.repositories.ProductRepository;
+import com.devjulio.dscommerce.services.exceptions.DatabaseException;
 import com.devjulio.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -51,18 +55,32 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id,ProductDTO dto){
-        Product entity = productRepository.getReferenceById(id);
+        try {
+            Product entity = productRepository.getReferenceById(id);
 
-        copyDtoToEntity(dto, entity);
+            copyDtoToEntity(dto, entity);
 
-        entity = productRepository.save(entity);
+            entity = productRepository.save(entity);
 
-        return new ProductDTO(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
-    @Transactional
-    public void deleteById(Long id){
-        productRepository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)// so executa a transação se este metodo estiver no contexto de outra transação
+    public void delete(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
